requirements.txt
aiogram==2.25.1
bs4==0.0.1
environs==9.5.0
asyncpg==0.27.0
forex-python==1.8
selenium==4.11.2
SQLAlchemy~=2.0.19
SQLAlchemy-Utils==0.41.1
python-dotenv==1.0.0
psycopg2-binary==2.9.7
aioredis==2.0.1
APScheduler==3.10.1
redis==5.0.0
python-slugify==8.0.1
